Refactor Java backend packages
