+++ backend-java/src/main/java/com/example/demo/model/UserEntity.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
