+++ backend-java/src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,7 +1,9 @@
-package com.example.demo;
+package com.example.demo.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.demo.model.UserEntity;
 
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByEmail(String email);
